+++ include/mgmt/mcumgr/smp.h
