+++ include/mgmt/mcumgr/smp_bt.h
