+++ subsys/CMakeLists.txt
@@ -8,7 +8,7 @@ add_subdirectory_ifdef(CONFIG_SHELL                shell)
 add_subdirectory_ifdef(CONFIG_CPLUSPLUS            cpp)
 add_subdirectory_ifdef(CONFIG_DISK_ACCESS          disk)
 add_subdirectory(fs)
-add_subdirectory_ifdef(CONFIG_MCUMGR               mgmt)
+add_subdirectory(mgmt)
 add_subdirectory_ifdef(CONFIG_MCUBOOT_IMG_MANAGER  dfu)
 add_subdirectory_ifdef(CONFIG_NET_BUF              net)
 add_subdirectory_ifdef(CONFIG_USB                  usb)
